Implement VScan models
One problem with various LLM providers is they can use arbitrary identifiers to call the individual LLMs. Even if they use the same model under the hood, for instance, GPT 4o is identified by OpenAI as gpt-4o, OpenRouter as openai/gpt-4o, and nano-gpt uses gpt-4o just like OpenAI, however, for the 4.1 generation of the GPT models, they use another tagging - openai/gpt-4.1, while OpenAI uses just gpt-4.1.
VScan uses a simple single string identifier for assigning models to configs. This however is not sufficient, because as can be seen on the mentioned example, any identifier would be tightly coupled to the specific provider and wouldn't work if the user decided to swap the provider of that particular model for another-one.

To resolve this situation, in this commit, VScan comes with its own model identifiers for the supported models. Simply put, VScan itself has a built-in catalogue of common vision language models, which use VScan's own textual identifiers. These always start with vscan-, e.g. vscan-gpt-4o, vscan-claude-4-sonnet etc.
Each provider created by the user now includes a list of supported models, where the VScan model ID is assigned to the ID used by the respective provider. I.E. when creating a provider representing OpenAI, vscan-gpt-4o points to gpt-4o. A provider for OpenRouter would point vscan-gpt-4o to openai/gpt-4o. And so on.
The provider presets should include the mappings of all supported models from the catalogue, so by default, the user won't need to configure anything if they're going with one of the common options.
At the same time, this provides a full variability on the providers used for models, because now when the user swaps the provider configured for a particular VScan known model, the according provider model ID will be used automatically without a need to change Configs, which thus stay universal.
However, in case the user wants to use a model not supported by the catalogue, they should be able to. Any model identifier that doesn't start with vscan- will be looked up among the provider models like a VScan identifier, but if no mapping is found, it will simply be used as is. If the model identifier does start with vscan- but is not found among the provider models, an error is shown, because VScan can not determine the identifier to be used for the particular provider.
The Vscan model catalogue as well as the mappings for individual providers should be expanded and updated over time.