Add support for capture modes
The default MAXIMIZE_QUALITY capture mode of ImageCapture has a quite long latency, about 600 - 700 ms, which prolonges an already relatively long waiting time on the model to return response. While for many use-cases, the gained image quality may not actually be necessary.

This commit starts the work of implementing different image obtaining methods, represented by the new Config's attribute. The user can now specify the capture mode option passed to the ImageCapture used for taking images, but also choose a video option, which launches a video stream in the background and upon request returns the next available image. This reduces the latency to 30 - 100 ms, depending on various factors.

The functionality is pretty much implemented and works, one notable exception is the flashlight, which uses a different approach for ImageCapture than it does for ImageAnalysis.